Dev Small fixes

+++ application/controllers/admin/ParticipantsAction.php
@@ -2502,7 +2502,7 @@ class ParticipantsAction extends SurveyCommonAction
         $participantIdsString = Yii::app()->request->getPost('participant_id'); // TODO: This is a comma separated string of ids
         $participantIds = explode(",", $participantIdsString);
 
-        $surveyId = Yii::app()->request->getPost('surveyid');
+        $surveyId = (int)Yii::app()->request->getPost('surveyid');
 
         /**
          * mapped can take values like
